session-api final refactor
